Eliminação da InterfaceEmpresa.java Finalização da Classe TipologiaCafe.java Remoção do atributo clientes diário da CategoriaRestauracao.java

+++ modulos/empresas/TipologiaFrutaria.java
@@ -11,12 +11,12 @@ public class TipologiaFrutaria extends CategoriaMercearia {
     }
 
     @Override
-    public int calcularDespesaAnual() {
+    public float calcularDespesaAnual() {
         return 0;
     }
 
     @Override
-    public int calcularReceitaAnual() {
+    public float calcularReceitaAnual() {
         return 0;
     }
 }
