+++ modulos/empresas/TipologiaMercado.java
@@ -15,12 +15,12 @@ public class TipologiaMercado extends CategoriaMercearia{
     }
 
     @Override
-    public int calcularDespesaAnual() {
+    public float calcularDespesaAnual() {
         return 0;
     }
 
     @Override
-    public int calcularReceitaAnual() {
+    public float calcularReceitaAnual() {
         return 0;
     }
 }
